+++ tests/test_utils.py
@@ -22,43 +22,13 @@ def mock_validated(mocker):
 
 
 @pytest.fixture
-def mock_download(mocker):
-    return mocker.patch.object(utils, 'download_from_remote')
-
-
-@pytest.fixture
-def mock_untar(mocker):
-    return mocker.patch.object(utils, 'untar')
-
-
-@pytest.fixture
-def mock_unzip(mocker):
-    return mocker.patch.object(utils, 'unzip')
-
-
-@pytest.fixture
 def mock_validator(mocker):
     return mocker.patch.object(utils, 'validator')
 
 
 @pytest.fixture
-def mock_force_delete_all(mocker):
-    return mocker.patch.object(utils, 'force_delete_all')
-
-
-@pytest.fixture
 def mock_check_index(mocker):
     return mocker.patch.object(utils, 'check_index')
-
-
-@pytest.fixture
-def mock_create_invalid(mocker):
-    return mocker.patch.object(utils, 'create_invalid')
-
-
-@pytest.fixture
-def mock_create_validated(mocker):
-    return mocker.patch.object(utils, 'create_validated')
 
 
 def test_md5(mocker):
@@ -113,100 +83,3 @@ def test_validator(mocker,
 
 def test_get_default_dataset_path():
     assert '/tmp/mir_datasets/data_home' == utils.get_default_dataset_path('data_home')
-
-
-def test_download_from_remote(httpserver, tmpdir):
-    httpserver.serve_content(open('tests/resources/remote.wav').read())
-
-    TEST_REMOTE = utils.RemoteFileMetadata(
-        filename='remote.wav',
-        url=httpserver.url,
-        checksum=('3f77d0d69dc41b3696f074ad6bf2852f')
-    )
-
-    download_path = utils.download_from_remote(TEST_REMOTE, str(tmpdir))
-    expected_download_path = os.path.join(str(tmpdir), 'remote.wav')
-    assert expected_download_path == download_path
-
-
-def test_download_from_remote_raises_IOError(httpserver, tmpdir):
-    httpserver.serve_content('File not found!', 404)
-
-    TEST_REMOTE = utils.RemoteFileMetadata(
-        filename='remote.wav',
-        url=httpserver.url,
-        checksum=('1234')
-    )
-
-    with pytest.raises(IOError):
-        utils.download_from_remote(TEST_REMOTE, str(tmpdir))
-
-
-def test_unzip(tmpdir):
-    utils.unzip('tests/resources/remote.zip', str(tmpdir))
-
-    expected_file_location = os.path.join(str(tmpdir), 'remote.wav')
-    assert os.path.exists(expected_file_location)
-
-
-def test_untar(tmpdir):
-    utils.untar('tests/resources/remote.tar.gz', str(tmpdir))
-
-    expected_file_location = os.path.join(str(tmpdir), 'remote.wav')
-    assert os.path.exists(expected_file_location)
-
-
-# def test_check_validated(tmpdir):
-#     tmpdir_str = str(tmpdir)
-#     assert not utils.check_validated(tmpdir_str)
-
-#     utils.create_validated(tmpdir_str)
-#     assert utils.check_validated(tmpdir_str)
-
-
-# def test_create_validated(tmpdir):
-#     tmpdir_str = str(tmpdir)
-#     expected_validated_path = os.path.join(tmpdir_str, utils.VALIDATED_FILE_NAME)
-#     assert not os.path.exists(expected_validated_path)
-
-#     utils.create_validated(tmpdir_str)
-#     assert os.path.exists(expected_validated_path)
-#     with open(expected_validated_path, 'r') as f:
-#         # Yes we could do not f.read(), but the intentions here are clearer
-#         assert f.read() == ''
-
-
-# @pytest.mark.parametrize('missing_files,invalid_checksums', [
-#     ({'10161_chorus': ['tests/resources/10162_chorus.wav']}, {}),
-#     ({}, {'10161_chorus': ['tests/resources/10161_chorus.wav']}),
-# ])
-# def test_create_invalid(tmpdir,
-#                         missing_files,
-#                         invalid_checksums):
-#     tmpdir_str = str(tmpdir)
-#     utils.create_invalid(tmpdir_str, missing_files, invalid_checksums)
-
-#     with open(os.path.join(tmpdir_str, utils.INVALID_FILE_NAME)) as f:
-#         invalid_content = json.load(f)
-#         assert invalid_content == {'missing_files': missing_files,
-#                                    'invalid_checksums': invalid_checksums}
-
-
-def test_force_delete_all_nonempty_data_home(httpserver, tmpdir):
-    tmpdir_str = str(tmpdir)
-    remote_filename = 'remote.wav'
-    TEST_REMOTE = utils.RemoteFileMetadata(
-        filename=remote_filename,
-        url=httpserver.url,
-        checksum=('1234')
-    )
-
-    with pytest.raises(IOError):
-        utils.download_from_remote(TEST_REMOTE, tmpdir_str)
-
-    utils.untar('tests/resources/remote.tar.gz', tmpdir_str)
-    assert os.path.exists(os.path.join(tmpdir_str, remote_filename))
-    assert os.path.exists(tmpdir_str)
-    utils.force_delete_all(TEST_REMOTE, tmpdir_str)
-    assert not os.path.exists(os.path.join(tmpdir_str, remote_filename))
-    assert not os.path.exists(tmpdir_str)
